+++ src/modules/aborter/aborter.test.ts
@@ -133,6 +133,15 @@ describe('Aborter', () => {
 
       expect(timeoutMock.clearTimeout).toHaveBeenCalled();
     });
+
+    it('Свойство isAborted должно быть true', () => {
+      aborter['isRequestInProgress'] = true;
+
+      aborter.abort('test reason');
+
+      expect(aborter['isRequestInProgress']).toBe(false);
+      expect(aborter.isAborted).toBeTruthy();
+    });
   });
 
   describe('Метод abortWithRecovery', () => {